Add puppet_openstack_builder to puppet-group
Puppet_openstack_builder is a composiiton layer for Puppetized
deployments of OpenStack that was formally introduced into the
StackForge community in the last development cycle.  Today, it
is not included in the puppet-group in Stackalytics even though
it is managed by the same group of core devs.  This patch adds
puppet_openstack_builder to the "puppet-group" module group.

Change-Id: I509deb8a8a80742b682d34ef81b634a21997b20b