Charter purchases Time Warner Cable
"Charter Communications won approval from California regulators to buy
Time Warner Cable and create the nation's second-largest cable company,
Bloomberg reports."

The deal closes May 18, 2016.

Note: Stackalytics requires us to keep the old company in the list as
long as our users are attached to it. In order to prevent users with not
entry from catching the twcable.com domain and being assigned into a
non-existant company, we'll use TWC's old domain which is unused for
email.

Change-Id: Iba9a09a67c972fe5ccad35bdd568cde75839cdab