Update info for Elise Gafford
Updating secondary to name change. Please review the logic in
record_processor.update_user: I believe that per this logic,
if all sources of authority (Gerrit, Zanata, Launchpad) have
already been scraped and stored with the user, user name
will never be updated if it is changed in those sources of
authority.

Stackalytics is a fairly minor service to require an additional
manual update in this case, and could easily be overlooked. In
some cases, this could pose a personal difficulty for those
unaware of the persistent name mismatch.

It's quite possible that I'm misreading the code, that there's
an additional periodic refresh cycle, or somesuch. If so,
thanks for already thinking of this. :)

If this is a possible issue, please do let me know; I'm happy
to file the bug (and even fix it if it's deemed low priority.)
Thanks!

Change-Id: Ia91d84c1f3484d3509da1550113c06f3ada75438