Hide all deb-* projects (except "deb-auto-backports")
Master branch of deb-* repos contains commits from upstream Debian
community and should not be calculated as OpenStack's.
OpenStack-specific contribution is done in debian/<release> branch.
But unlike in the rest of OpenStack, git workflow differs and
the branch contains merge commits from master. This makes
filtering "pure" branch commits from those that came from master
quite tricky (not possible to specify the branch-point).
And support of this will require changes in Stackalytics code.

With this commit all deb-* repos are hidden from stats until
Stackalytics is extended to support the alternative workflow.

Notification regarding this change is sent in:
http://lists.openstack.org/pipermail/openstack-dev/2016-September/104139.html

Change-Id: Ie1a2dac7d4f02070f3a54bb1afea6207bd6692b0