Add destructor to SSHClient
Newer versions of paramiko require a client object to be explicitly
closed.  Fortunately, we wrap all of our use of paramiko client
objects in our own class.  Add a destructor to our class which
closes the client object.

Note, this has been tested to work (and is needed) even if a
connection is not established.

Change-Id: Iaadc8efc84efa551b0fa659b4b9b19dca29edf43
Signed-off-by: Paul Belanger <pabelanger@redhat.com>