Revert "Add support for user correction"
This reverts commit bac722a749b11ff0d7cd49b67d7b75ac73658023.

There are 2 user profiles that mix with each other because of the same
email re-used by other person. The proposed change was intended to fix
this, but it does not. All it does is update of user profile after
record processing, but at that moment records already refer to
the wrong person. The correct solution is to somehow restrict usage
of email (e.g. by date) or at least restrict mapping of particular
user profiles.

Change-Id: I1c37bd9778d740d5095f08885bb3cb63a656a604