Specify user_id on load_user() calls
Two load_user() calls didn't specify the argument user_id in
record_processor module, that meaned the parameters were passed
as seq, instead of user_id.
In load_user(), seq and user_id are handled as the same way and
we didn't face any problem due to this bug. This patch makes
the method specify the argument clearly to avoid issues when
changing internal behavior of load_user().

Change-Id: I1bfd08c273bd0272eada7cb3b27930a0f7a43315