Update e-mail addresses for jsbryant
I have noticed that, despite being more active
on the mailing lists, I have not seen this reflected
in Stackalytics.

Further review shows that this is probably because my
e-mails are showing up as having GMail as the source
even though the reply-to is my electronicjungle.net
accounts.

I am adding the GMail accounts with this patch.

Change-Id: I0b8b49c4fd467afc6600eb97ed67bc991e2f5e97