Adding OpenLab
Please see http://openlabtesting.org and
https://github.com/theopenlab

OpenLab’s mission is to enable the testing, reporting, and development
of tools and applications for hybrid and multi-cloud environments.
OpenLab fulfills its mission by providing a ready to use and easy to
replicate platform that tests the usability, reliability, and resiliency
of hybrid and multi-cloud tooling.

Change-Id: I887b08c4a5f8bdf7b4e877d550a2f90a4b3059b5