In timeline hide the second line if it has only zeros
Better late than never - no more misleading second y axis in
the timeline.

The second line is used only for commits and LOCs metrics.
For the rest it appears on chart as flat yellow line and y axes
showing 0, 0.5 and 1.0. With this patch the timeline becomes clearer
and does not show zero line anymore.

Change-Id: I28072204d57088e222697cb6bd176bd8769938ad